Sigo con DatosCliente.cs

+++ Domain/Negocio/Cliente.cs
@@ -13,6 +13,6 @@ namespace Domain.Negocio
         public string Direccion { get; set; }
         public string Localidad { get; set; }
         public string Provincia { get; set; }
-        public int Tarjeta { get; set; }
+        public long Tarjeta { get; set; }
     }
 }
